requirements.txt
django == 2.2.2
djangorestframework == 3.9.2
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0